backend/requirements.txt
# requirements.txt
flask
python-dotenv
supabase
flask-cors
yfinance